debug: Using csg rotate and translate
Weirdly this change and rotate-before-translation-and-then-extrude needed
this change to even partially render text. Without this there was no text
and, in fact in rotate-before-translation-and-then-extrude there polygon 2
was missing.

With this change the rotation doesn't work and we end up with the text
on the bottom inside the cube and thus not normally visible. But if you
use --opacity and --translucency-support you can just see the text:
  f3d --opacity=0.6 --translucency-support cube.len_side-2.397_tube_diameter-0.595_segments-50.stl &